output_to_fits moved to higher abstract of structure
